+++ src/config/frontendRoutes.ts
@@ -15,6 +15,7 @@ export enum FrontendRoute {
   RESET_PASSWORD_PAGE = '/password/reset',
   SPONSOR_SEARCH_PAGE = '/sponsor/search',
   TEAM_PAGE = '/team',
+  TRAVEL_PAGE = '/travel',
 
   CREATE_SPONSOR_PAGE = '/sponsor/create',
   EDIT_SPONSOR_PAGE = '/sponsor/edit',