+++ src/util/UserInfoHelperFunctions.tsx
@@ -116,14 +116,34 @@ export function canAccessTeam(hacker?: IHacker): boolean {
   );
 }
 
+export function canAccessTravel(hacker?: IHacker): boolean {
+  const status = hacker ? hacker.status : HackerStatus.HACKER_STATUS_NONE;
+
+  if (
+    status === HackerStatus.HACKER_STATUS_APPLIED ||
+    status === HackerStatus.HACKER_STATUS_ACCEPTED ||
+    status === HackerStatus.HACKER_STATUS_CONFIRMED ||
+    status === HackerStatus.HACKER_STATUS_CHECKED_IN
+  ) {
+    return !!(
+      hacker &&
+      hacker.application &&
+      hacker.application.accommodation &&
+      hacker.application.accommodation.travel &&
+      hacker.application.accommodation.travel > 0
+    );
+  }
+  return false;
+}
+
 export function canAccessBus(hacker?: IHacker): boolean {
   const status = hacker ? hacker.status : HackerStatus.HACKER_STATUS_NONE;
   return hacker
     ? Boolean(hacker.travel) &&
-        (status === HackerStatus.HACKER_STATUS_APPLIED ||
-          status === HackerStatus.HACKER_STATUS_ACCEPTED ||
-          status === HackerStatus.HACKER_STATUS_CONFIRMED ||
-          status === HackerStatus.HACKER_STATUS_CHECKED_IN)
+    (status === HackerStatus.HACKER_STATUS_APPLIED ||
+      status === HackerStatus.HACKER_STATUS_ACCEPTED ||
+      status === HackerStatus.HACKER_STATUS_CONFIRMED ||
+      status === HackerStatus.HACKER_STATUS_CHECKED_IN)
     : false;
 }
 
@@ -147,7 +167,7 @@ export async function generateHackerQRCode(hacker: IHacker): Promise<string> {
   const hackerPage = `
   ${window.location.protocol}//${window.location.hostname}${
     window.location.port ? ':' + window.location.port : ''
-  }${FrontendRoute.VIEW_HACKER_PAGE.replace(':id', hacker.id)}`;
+    }${FrontendRoute.VIEW_HACKER_PAGE.replace(':id', hacker.id)}`;
   const response = await QRCode.toDataURL(hackerPage, { scale: 10 });
   return response;
 }