+++ docs/CHANGELOG.md
@@ -5,13 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
-
 ### Added
 
 - Status page displays appropriate message based on hacker status
 - Job interest column for sponsors to see more info on hacker
+<<<<<<< HEAD
 - Added buttons for status page based on status
+=======
+- Added travel page that displays hacker's travel status and reimbursement amount
+>>>>>>> develop
 
 ### Fixed